Move source over from TFS
This project has languished in TFS for a couple years. Dust it off and
get it on GitHub so I'm more likely to guilt myself into poking around
on it.

This is missing the Minecraft import world save which
SimCityToMinecraft\inputModelConfig.xml refers to by coordinates. It
contains road / structure templates used to build the generated city
from.